Objective: Closes UI when Eaten during task
Corrected objectiveTask script to close UI when Player is eaten
during the task, and also make the OpenTaskButton disappear.
Deleted the SetTaskActive & Inactive funcs because it could be
done with the feature of the Button component.
Set Alien objs inactive (unchecked on the left of the name in the
inspector) to be able to test above.

resolving #20