alienMovement: rudamentary close quartre mechanic
When alien close to player, provided noise is below noiseThreshhold,
the alien will not know where the player is and will search for him.
This will happen over time indicated by circleLimit. Alien will stand
still for 1/3 of that time, circle for next 1/3, then circle wider and
faster for latter 1/3.

resolving #33